// src/app/Interfaces/PostalCode.ts
export interface postalCodeI {
    id: number
    cityId: number
    cityName: any
    stateId: number
    stateName: any
    countryId: number
    countryName: any
    postalCode: string
    latitude: string
    longitude: string
    isTaxApplicable: boolean
    taxRate: any
    isActive: boolean
    description: string
    lastActionBy: any
    lastActionOn: string
  }
  